fix(filter-drp): move FieldValidation type to FilterDRP
