+++ packages/date-picker/src/FilterDateRangePicker/hooks/useRangeDateValidation.ts
@@ -1,7 +1,8 @@
 import { useState } from "react"
 import { FieldMessageStatus } from "@kaizen/draft-form"
-import { FieldValidation, Matcher, ValidationResponse } from "../../types"
+import { Matcher, ValidationResponse } from "../../types"
 import { validateDate } from "../../utils/validateDate"
+import { FieldValidation } from "../types"
 import { getDateValidationHandler } from "../utils/getDateValidationHandler"
 
 export type UseRangeDateValidationArgs = {
