+++ packages/date-picker/src/FilterDateRangePicker/hooks/useStartDateValidation.ts
@@ -1,4 +1,4 @@
-import { FieldValidation } from "../../types"
+import { FieldValidation } from "../types"
 import {
   useRangeDateValidation,
   UseRangeDateValidationArgs,
